refactor: decompose the apply-job modal into a generation hook and subcomponents
Extract the analyze→resume→cover-letter flow and its state into
useTailorGeneration, and split the output panel (GenerationOutput) and the
job-description block (JobDescriptionPanel) into their own files. index.tsx is
now layout only.
